+++ source/DataAssimilator.cc
@@ -213,7 +213,8 @@ DataAssimilator::apply_kalman_gain(
   // members in parallel
   std::transform(
       perturbed_innovation.begin(), perturbed_innovation.end(), output.begin(),
-      [&](dealii::Vector<double> entry) {
+      [&](dealii::Vector<double> entry)
+      {
         dealii::SolverGMRES<dealii::Vector<double>> HPH_plus_R_inv_solver(
             solver_control, additional_data);
