+++ tests/data/experimental_data_0_0.csv
@@ -1,10 +1,10 @@
-"x", "y", "z", "a", "T", "b"
-0,0,1,0,1.2,0
-0,0.5,1,0,1.3,0
-0,1,1,0,1.4,0
-0.5,0,1,0,1.5,0
-0.5,0.5,1,0,1.6,0
-0.5,1,1,0,1.7,0
-1,0,1,0,1.8,0
-1,0.5,1,0,1.9,0
-1,1,1,0,2.,0
+"x", "y", "z", "T"
+0,0,1,1.2
+0,0.5,1,1.3
+0,1,1,1.4
+0.5,0,1,1.5
+0.5,0.5,1,1.6
+0.5,1,1,1.7
+1,0,1,1.8
+1,0.5,1,1.9
+1,1,1,2.
